Move SignIn/BioAuth component into a seperate folder
